requirements.txt
numpy==1.7.0
pyzmq==13.0.0
